Higgs tree maker

+++ NtupleMaker/scripts/checkCMS3NtupleProduction.sh
@@ -37,57 +37,73 @@ for f in $(find $chkdir -name condor.sub); do
   let countOK=0
   let dirok=1
   let nsubjobs=0
-  for logfilename in $(ls $d/Logs | grep -e "log_"); do
+  let nRunningJobs=0
+  for joblog in $(ls $d | grep -e ".log"); do
+    jobnumber=${joblog//".log"}
+    logfilename="log_job.${jobnumber}.txt"
+
+    let nsubjobs=$nsubjobs+1
+
     resb=( )
     rese=( )
     resf=( )
     ress=( )
 
-    jobnumber=${logfilename//log_job.}
-    jobnumber=${jobnumber//.txt}
-    runningjob=( )
+    let job_is_running=0
     for jobid in "${runningjobs[@]}"; do
       if [[ "${jobid}" == "${jobnumber}" ]]; then
-        runningjob+=( "${jobnumber}" )
+        let job_is_running=1
         break
       fi
     done
     
+    if [[ ${job_is_running} -eq 1 ]]; then
+      echo "$d is still running"
+      let nUNKNOWN=$nUNKNOWN+1
+      let dirok=0
+      let nRunningJobs=${nRunningJobs}+1
+      continue
+    fi
+
     fread="$d/Logs/$logfilename"
 
-    if [[ $skiplongfile -eq 1 ]]; then
-      let freadsize=$(stat --format=%s $fread)
-      if [[ $freadsize -gt 1000000 ]]; then
-        echo "Skipping $fread because its size = $freadsize > 1000000"
-        continue
+    if [[ -e ${fread} ]]; then
+      if [[ $skiplongfile -eq 1 ]]; then
+        let freadsize=$(stat --format=%s $fread)
+        if [[ $freadsize -gt 1000000 ]]; then
+          echo "Skipping $fread because its size = $freadsize > 1000000"
+          continue
+        fi
+        #echo "File $fread with size $freadsize is not long..."
       fi
-      #echo "File $fread with size $freadsize is not long..."
     fi
 
-    let nsubjobs=$nsubjobs+1
-
-    if [[ $islongfile -eq 1 ]]; then
-      tmpfile="tail_$logfilename"
-      echo "Truncating $fread to $tmpfile"
-      tail -1000 $fread &> $tmpfile
-      fread=$tmpfile
-      echo "Will read $fread"
+    if [[ -e ${fread} ]]; then
+      if [[ $islongfile -eq 1 ]]; then
+        tmpfile="tail_$logfilename"
+        echo "Truncating $fread to $tmpfile"
+        tail -1000 $fread &> $tmpfile
+        fread=$tmpfile
+        echo "Will read $fread"
+      fi
     fi
 
-    while IFS='' read -r line || [[ -n "$line" ]]; do
-      if [[ "$line" == *"begin copying output"* ]]; then
-        resb+=( "$line" )
-      elif [[ "$line" == *"end copying output"* ]]; then
-        rese+=( "$line" )
-      elif [[ "$line" == *"Running: env -i "* ]]; then
-        resf+=( "$line" )
-      elif [[ "$line" == *"Copied successfully"* ]]; then
-        ress+=( "$line" )
-      fi
-    done < "$fread"
+    if [[ -e ${fread} ]]; then
+      while IFS='' read -r line || [[ -n "$line" ]]; do
+        if [[ "$line" == *"begin copying output"* ]]; then
+          resb+=( "$line" )
+        elif [[ "$line" == *"end copying output"* ]]; then
+          rese+=( "$line" )
+        elif [[ "$line" == *"Running: env -i "* ]]; then
+          resf+=( "$line" )
+        elif [[ "$line" == *"Copied successfully"* ]]; then
+          ress+=( "$line" )
+        fi
+      done < "$fread"
+    fi
 
     if [[ $islongfile -eq 1 ]]; then
-      rm $fread
+      rm -f $fread
     fi
 
     let size_resb=${#resb[@]}
@@ -95,11 +111,7 @@ for f in $(find $chkdir -name condor.sub); do
     let size_resf=${#resf[@]}
     let size_ress=${#ress[@]}
 
-    if [[ ${#runningjob[@]} -gt 0 ]] && [[ ${runningjob[0]} != "" ]]; then
-      echo "$d is still running"
-      let nUNKNOWN=$nUNKNOWN+1
-      let dirok=0
-    elif [[ $size_resb -gt 0 ]] && [[ $size_resb -eq $size_rese ]] && [[ $size_resb -eq $size_ress ]] && [[ $size_resb -eq $size_resf ]];then
+    if [[ $size_resb -gt 0 ]] && [[ $size_resb -eq $size_rese ]] && [[ $size_resb -eq $size_ress ]] && [[ $size_resb -eq $size_resf ]];then
       let nOutputExist=0
       for rf in "${resf[@]}";do
         rf="${rf//*'gsiftp://gftp.t2.ucsd.edu'}"
@@ -109,16 +121,16 @@ for f in $(find $chkdir -name condor.sub); do
         fi
       done
       if [[ $nOutputExist -eq $size_resf ]];then
-        echo "$d ran successfully with $nOutputExist files."
+        echo "Job $jobnumber for $d ran successfully with $nOutputExist files."
         let nOK=$nOK+1
         let countOK=$countOK+1
       else
-        echo "$d ran successfully, but some files do not exist! (Nbegin, Ncopyrun, Nsuccess, Nend, Nexists) = ( $size_resb, $size_resf, $size_ress, $size_rese, $nOutputExist )"
+        echo "Job $jobnumber for $d ran successfully, but some files do not exist. (Nbegin, Ncopyrun, Nsuccess, Nend, Nexists) = ( $size_resb, $size_resf, $size_ress, $size_rese, $nOutputExist )"
         let nFILEDNE=$nFILEDNE+1
         let dirok=0
       fi
     else
-      echo "$d failed. (Nbegin, Ncopyrun, Nsuccess, Nend) = ( $size_resb, $size_resf, $size_ress, $size_rese )"
+      echo "Job $jobnumber for $d did not run successfully. (Nbegin, Ncopyrun, Nsuccess, Nend) = ( $size_resb, $size_resf, $size_ress, $size_rese )"
       let nFAIL=$nFAIL+1
       let dirok=0
     fi
@@ -127,7 +139,7 @@ for f in $(find $chkdir -name condor.sub); do
 
   if [[ $countOK -gt 0 ]] && [[ $dirok -eq 0 ]];then
     if [[ $multiprod -eq 1 ]];then
-      echo "$d has multiple submissions with $countOK / $nsubjobs success rate, but the folder will be treated as if it failed."
+      echo "$d has multiple submissions with $countOK / $nsubjobs success rate, but all subjobs were required to succeed."
     else
       echo "$d has multiple submissions with $countOK / $nsubjobs success rate, but the user specified the folders to be treated as single jobs."
       let dirok=1
@@ -135,19 +147,28 @@ for f in $(find $chkdir -name condor.sub); do
   fi
 
   if [[ $nsubjobs -eq 0 ]];then
-    echo "$d does not have any subjobs run yet. Marking as failed."
+    echo "$d does not have any subjobs run yet."
     let nFAIL=$nFAIL+1
     let dirok=0
   fi
 
   if [[ $dirok -eq 1 ]];then
     TARFILE="${d}.tar"
+    TARFILEFirst=${TARFILE%/*}
+    TARFILELast=${TARFILE##*/}
+    if [[ ${#TARFILELast} -gt 255 ]]; then
+      TARFILE=${TARFILEFirst}/${TARFILELast//_}
+    fi
     rm -f $TARFILE
     tar Jcf ${TARFILE} $d --exclude={*.tar}
     if [[ $? -eq 0 ]];then
       echo "- Compressed successfully, so removing the directory"
       rm -rf $d
     fi
+  elif [[ $nRunningJobs -gt 0 ]]; then
+    echo "$d is still running."
+  else
+    echo "$d failed."
   fi
 
 done
